requirements.txt
streamlit==1.31.0
torch==2.2.0
Pillow==10.2.0
requests==2.31.0
transformers==4.38.0
matplotlib==3.8.2
pandas==2.2.0
base64==1.0.0